+++ src/organisms/headings/carousel/carousel.vue
@@ -21,33 +21,25 @@
       </div>
 
       <div class="osg-carousel__navigation" v-if="this.images.length > 1">
-        <a
-          @click.prevent="goToPrev"
-          target="_self"
-          href="#"
-        >
-          <osg-vue-button
+        <osg-vue-button
             class="osg-carousel__previous-button"
             :is-circle="true"
             :color="navigationArrowColor"
-          >
-            <osg-vue-icon :iconName="icons.previousIcon" />
-          </osg-vue-button>
-        </a>
-
-        <a
-          @click.prevent="goToNext"
-          target="_self"
-          href="#"
+            aria-label="Forrige bilde"
+            :on-click="goToPrev"
         >
-          <osg-vue-button
+          <osg-vue-icon :iconName="icons.previousIcon" />
+        </osg-vue-button>
+
+        <osg-vue-button
             class="osg-carousel__next-button"
             :is-circle="true"
             :color="navigationArrowColor"
-          >
-            <osg-vue-icon :iconName="icons.nextIcon" />
-          </osg-vue-button>
-        </a>
+            aria-label="Neste bilde"
+            :on-click="goToNext"
+        >
+          <osg-vue-icon :iconName="icons.nextIcon" />
+        </osg-vue-button>
       </div>
 
       <div class="osg-carousel__shapes">
@@ -55,7 +47,7 @@
           v-if="hasSquaredShape"
           :class="[
             'osg-carousel__squared-shape',
-            `osg-u-color-bg-${squareColor}` 
+            `osg-u-color-bg-${squareColor}`
           ]"
         />
 
@@ -128,7 +120,7 @@
         type: Object,
         default: () => null
       },
-    
+
       icons: {
         type: Object,
         default () {
@@ -138,7 +130,7 @@
           }
         }
       },
-      
+
       images: {
         type: Array,
         required: true
@@ -282,7 +274,7 @@
           slide.classList.add("osg-carousel__slide")
         }
       },
-      
+
       /**
        * Prepare carousel styles
        */