+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - molecules-forms-input_date: adds component
 
+- molecules-forms-input-datepicker: adds component
+
 ### Removed
 
 ### Changed