+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - atoms-icons-icon: fix A11y, adds aria-hidden
 - atoms-icons-icon: fix BEM block class
+- organisms-headings-carousel: fix A11y
 
 ### Added
 - molecules-forms-input_date: adds component